+++ src/components/Layout/Modal/Modal.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, cleanup, fireEvent } from '@testing-library/react';
-import Modal from '.';
+import Modal from './index';
 
 describe('Modal component', () => {
   afterEach(() => {