+++ src/components/Typography/Link/Link.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import { Color } from '../../../shared';
@@ -6,21 +6,19 @@ import { Color } from '../../../shared';
 import './link.scss';
 
 /**
- * Beautiful-ui does not force the developer to use its own styles nor creates extra global rules that can possibly
- * collide with the application's ones.<br/>
- * For this reason, in order to possibly have the same style between UI components and texts when needed, few
- * typography specialised components has been created.<br/>
+ * beautiful-react-ui does not force the developer to use its own styles nor creates extra global rules that can
+ * possibly collide with the application's ones.<br/>
+ * For this reason, in order to have the consistent style between the UI components and the texts, few typography
+ * specialised components has been created.<br/>
  * The typography specialised components are used within the library itself.
  * <br/>
  * Here's the Link component.
  */
 const Link = (props) => {
-  const { href, color, children, className, ...rest } = props;
-  const classList = classNames('bi bi-link', `bi-link-${color}`, className);
+  const { href, color, children, className, ElementRender, ...rest } = props;
+  const classList = useMemo(() => classNames('bi bi-link', `bi-link-${color}`, className), [color, className]);
 
-  return (
-    <a href={href} className={classList} {...rest}>{children}</a>
-  );
+  return (<ElementRender href={href} className={classList} {...rest}>{children}</ElementRender>);
 };
 
 Link.propTypes = {
@@ -33,11 +31,16 @@ Link.propTypes = {
    * `default`, `primary`, `secondary`, `info`, `warning`, `success`, `danger`.
    */
   color: Color,
+  /**
+   * A renderer to replace the link element
+   */
+  ElementRender: PropTypes.elementType,
 };
 
 
 Link.defaultProps = {
   color: 'primary',
+  ElementRender: 'a',
 };
 
 export default React.memo(Link);