+++ src/components/Typography/Paragraph/Paragraph.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import { Color } from '../../../shared';
@@ -6,26 +6,24 @@ import { Color } from '../../../shared';
 import './paragraph.scss';
 
 /**
- * Beautiful-ui does not force the developer to use its own styles nor creates extra global rules that can possibly
- * collide with the application's ones.<br/>
- * For this reason, in order to possibly have the same style between UI components and texts when needed, few
- * typography specialised components has been created.<br/>
+ * beautiful-react-ui does not force the developer to use its own styles nor creates extra global rules that can
+ * possibly collide with the application's ones.<br/>
+ * For this reason, in order to have the consistent style between the UI components and the texts, few typography
+ * specialised components has been created.<br/>
  * The typography specialised components are used within the library itself.
  * <br/>
  * Here's the Paragraph component.
  */
 const Paragraph = (props) => {
-  const { children, color, fontFamily, textAlign, wordBreak, tiny, light, className, ...rest } = props;
-  const classList = classNames('bi bi-p', `bi-p-${color}`, `bi-ff-${fontFamily}`, {
+  const { children, color, fontFamily, textAlign, wordBreak, tiny, light, className, ElementRender, ...rest } = props;
+  const classList = useMemo(() => classNames('bi bi-p', `bi-p-${color}`, `bi-ff-${fontFamily}`, {
     [`bi-p-${textAlign}`]: !!textAlign,
     [`bi-p-break-${wordBreak}`]: !!wordBreak,
     'bi-p-tiny': tiny,
     'bi-p-light': light,
-  }, className);
+  }, className), [color, fontFamily, textAlign, wordBreak, tiny, light]);
 
-  return (
-    <p className={classList} {...rest}>{children}</p>
-  );
+  return (<ElementRender className={classList} {...rest}>{children}</ElementRender>);
 };
 
 Paragraph.propTypes = {
@@ -55,6 +53,10 @@ Paragraph.propTypes = {
    * Makes the text color lighter
    */
   light: PropTypes.bool,
+  /**
+   * A renderer to replace the paragraph element
+   */
+  ElementRender: PropTypes.elementType,
 };
 
 
@@ -65,6 +67,7 @@ Paragraph.defaultProps = {
   wordBreak: 'normal',
   tiny: false,
   light: false,
+  ElementRender: 'p',
 };
 
 export default React.memo(Paragraph);